service-workers/service-worker/ServiceWorkerGlobalScope/registration-attribute.https.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: registration</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../resources/test-helpers.sub.js'></script>
<script>

promise_test(async t => {
  t.add_cleanup(async() => {
    if (frame)
      frame.remove();
    if (registration)
      await registration.unregister();
  });

  const SCRIPT = 'resources/registration-attribute-worker.js';
  const SCOPE = 'resources/scope/registration-attribute';
  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);
  const expected_events_seen = [
    'updatefound',
    'install',
    'statechange(installed)',
    'statechange(activating)',
    'activate',
    'statechange(activated)',
    'fetch',
  ];
  assert_equals(
    frame.contentDocument.body.textContent,
    expected_events_seen.toString(),
    'Service Worker should respond to fetch'
  );
}, 'Verify registration attributes on ServiceWorkerGlobalScope');

promise_test(async t => {
 t.add_cleanup(async() => {
    if (registration1)
      await registration1.unregister();
    if (registration2)
      await registration2.unregister();
 });

  const SCRIPT1 = 'resources/registration-attribute-worker.js';
  const SCRIPT2 = 'resources/registration-attribute-newer-worker.js';
  const SCOPE = 'resources/scope/registration-attribute';
  const registration1 = await service_worker_unregister_and_register(
    t,
    SCRIPT1,
    SCOPE
  );
  await wait_for_state(t, registration1.installing, 'activated');
  const registration2 = await navigator.serviceWorker.register(
    SCRIPT2,
    {scope: SCOPE}
  );
  assert_equals(registration1, registration2);
  const worker = registration2.installing;
  await wait_for_state(t, worker, 'activated');
  const channel = new MessageChannel
  worker.postMessage({port: channel.port2}, [channel.port2]);
  const results = await new Promise(function(resolve) {
    channel.port1.onmessage = function(e) { resolve(e.data); };
  });
  const script_url = normalizeURL(SCRIPT1);
  const newer_script_url = normalizeURL(SCRIPT2);
  const expectations = [
    'evaluate',
    '  installing: empty',
    '  waiting: empty',
    '  active: ' + script_url,
    'updatefound',
    '  installing: ' + newer_script_url,
    '  waiting: empty',
    '  active: ' + script_url,
    'install',
    '  installing: ' + newer_script_url,
    '  waiting: empty',
    '  active: ' + script_url,
    'statechange(installed)',
    '  installing: empty',
    '  waiting: ' + newer_script_url,
    '  active: ' + script_url,
    'statechange(activating)',
    '  installing: empty',
    '  waiting: empty',
    '  active: ' + newer_script_url,
    'activate',
    '  installing: empty',
    '  waiting: empty',
    '  active: ' + newer_script_url,
    'statechange(activated)',
    '  installing: empty',
    '  waiting: empty',
    '  active: ' + newer_script_url,
  ];
  assert_array_equals(results, expectations);
}, 'Verify registration attributes on ServiceWorkerGlobalScope of the ' +
   'newer worker');

</script>
