third_party/WebKit/LayoutTests/platform/linux/virtual/service-worker-script-streaming/external/wpt/service-workers/service-worker/ServiceWorkerGlobalScope/registration-attribute.https-expected.txt

This is a testharness.js-based test.
FAIL Verify registration attribute on ServiceWorkerGlobalScope assert_equals: Service Worker should respond to fetch expected "updatefound,install,statechange(installed),statechange(activating),activate,statechange(activated),fetch" but got "updatefound,install,activate,statechange(installed),statechange(activating),statechange(activated),fetch"
Harness: the test ran to completion.
